+++ Realm/Swift/RLMSwiftSupport.swift
@@ -61,7 +61,7 @@ import Foundation
         // super is an implicit property on Swift objects
         for i in 1..<reflection.count {
             let propertyName = reflection[i].0
-            if ignoredPropertiesForClass?.containsObject(propertyName) {
+            if ignoredPropertiesForClass && ignoredPropertiesForClass!.containsObject(propertyName) {
                 continue
             }
 
@@ -86,7 +86,12 @@ import Foundation
             case is Bool.Type, is Bool?.Type:
                 (p, t) = (RLMProperty(name: name, type: .Bool, objectClassName: nil, attributes: attr), "c")
             case is Int.Type, is Int?.Type:
-                (p, t) = (RLMProperty(name: name, type: .Int, objectClassName: nil, attributes: attr), "i")
+                p = RLMProperty(name: name, type: .Int, objectClassName: nil, attributes: attr)
+#if arch(x86_64) || arch(arm64)
+                t = "l"
+#else
+                t = "i"
+#endif
             case is Float.Type, is Float?.Type:
                 (p, t) = (RLMProperty(name: name, type: .Float, objectClassName: nil, attributes: attr), "f")
             case is Double.Type, is Double?.Type: